Made a new Perm Check

+++ src/Commands/Admin/Database/DBReset.js
@@ -10,11 +10,11 @@ class DBDebug extends Command {
   constructor() {
     super("dbreset", {
       description: "Resets the guild's settings.",
-      typing: true,
-      userPermissions: ["ADMINISTRATOR"]
+      typing: true
     });
   }
   async exec(m) {
+    if (!m.member.permissions.has["ADMINISTRATOR"] && m.author.id != this.client.ownerID) return m.channel.send("You do not have valid permissions to run this command. You need the **Administrator** permission.");
     await this.client.mongo.guilds.deleteMany({ gid: m.guild.id });
     await this.client.mongo.createGuild(m.guild.id);
     return m.channel.send("Your guild's settings has been successfully reset.");
